Switch Next version and revert to JS _document
