+++ package.json
@@ -20,7 +20,7 @@
     "graphql-request": "^3.4.0",
     "gsap": "^3.7.0",
     "isomorphic-unfetch": "^3.1.0",
-    "next": "11.1.0",
+    "next": "^11.0.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-fast-compare": "^3.2.0",
@@ -35,8 +35,8 @@
     "use-persisted-state": "^0.3.3"
   },
   "devDependencies": {
-    "@types/node": "20.11.0",
-    "@types/react": "18.2.47",
+    "@types/node": "^20.11.0",
+    "@types/react": "^17.0.2",
     "typescript": "^5.3.3"
   }
 }
